requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
albumentations=1.1.0=pypi_0
async-timeout=4.0.2=pypi_0
attrs=21.4.0=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2022.3.29=haa95532_0
cachetools=4.2.4=pypi_0
certifi=2021.10.8=py38haa95532_2
cffi=1.15.0=py38h2bbff1b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.4=pypi_0
cryptography=36.0.0=py38h21b164f_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pypi_0
efficientnet-pytorch=0.7.1=pypi_0
et_xmlfile=1.1.0=py38haa95532_0
fonttools=4.32.0=pypi_0
freetype=2.10.4=hd328e21_0
frozenlist=1.3.0=pypi_0
fsspec=2022.3.0=pypi_0
future=0.18.2=pypi_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.44.0=pypi_0
idna=3.3=pyhd3eb1b0_0
imageio=2.16.2=pypi_0
importlib-metadata=4.11.3=pypi_0
intel-openmp=2021.4.0=haa95532_3556
joblib=1.1.0=pypi_0
jpeg=9d=h2bbff1b_0
kiwisolver=1.4.2=pypi_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
markdown=3.3.6=pypi_0
matplotlib=3.5.1=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
multidict=6.0.2=pypi_0
networkx=2.8=pypi_0
numpy=1.21.5=py38ha4e8547_0
numpy-base=1.21.5=py38hc2deb75_0
oauthlib=3.2.0=pypi_0
opencv-python=4.5.5.64=pypi_0
opencv-python-headless=4.5.5.64=pypi_0
openpyxl=3.0.9=pyhd3eb1b0_0
openssl=1.1.1n=h2bbff1b_0
packaging=21.3=pypi_0
pandas=1.4.2=pypi_0
pillow=9.0.1=py38hdc2b20a_0
pip=21.2.2=py38haa95532_0
protobuf=3.20.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.8=pypi_0
pysocks=1.7.1=py38haa95532_0
python=3.8.13=h6244533_0
python-dateutil=2.8.2=pypi_0
pytorch=1.11.0=py3.8_cuda11.3_cudnn8_0
pytorch-lightning=0.8.5=pypi_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pypi_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0=pypi_0
qudida=0.0.4=pypi_0
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-image=0.19.2=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
seaborn=0.11.2=pypi_0
setuptools=58.0.4=py38haa95532_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.2=h2bbff1b_0
tensorboard=2.2.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.4.8=pypi_0
tk=8.6.11=h2bbff1b_0
torchaudio=0.11.0=py38_cu113
torchvision=0.12.0=py38_cu113
tqdm=4.64.0=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.8=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.1.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38haa95532_2
xz=5.2.5=h62dcd97_0
yarl=1.7.2=pypi_0
zipp=3.8.0=pypi_0
zlib=1.2.11=hbd8134f_5
zstd=1.4.9=h19a0ad4_0
